LU-20373 build: Compatibility updates for kernel v7.1
Linux commit v7.0-rc6-4-g14a51045e10d3
  get rid of busy-waiting in shrink_dcache_tree()
struct dentry.d_alias is now a member of an anonymous union.
provide a wrapper for older kernels to use the older named d_u
union to access d_alias

Linux commit v7.0-rc5-146-g4e1d77a8f382a
  folio_batch: rename pagevec.h to folio_batch.h
Provide a compat folio_batch.h that includes pagevec.h for
older kernels.

Linux commit v7.0-rc1-1-g96fefcabf340f
  vfs: widen inode hash/lookup functions to u64
Force legacy i_ino to u64 and change format to %llu

Test-Parameters: trivial
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/66662
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>

+++ lustre/llite/llite_lib.c
@@ -3053,8 +3053,9 @@ int ll_update_inode(struct inode *inode, struct lustre_md *md)
 	if (body->mbo_valid & OBD_MD_FLMTIME) {
 		if (body->mbo_mtime > inode_get_mtime_sec(inode)) {
 			CDEBUG(D_INODE,
-			       "setting ino %lu mtime from %lld to %llu\n",
-			       inode->i_ino, (s64) inode_get_mtime_sec(inode),
+			       "setting ino %llu mtime from %lld to %llu\n",
+			       (u64)inode->i_ino,
+			       (s64)inode_get_mtime_sec(inode),
 			       body->mbo_mtime);
 			inode_set_mtime(inode, body->mbo_mtime, 0);
 		}
