+++ lustre/llite/vvp_object.c
@@ -47,8 +47,8 @@ static int vvp_object_print(const struct lu_env *env, void *cookie,
 	     inode);
 	if (inode) {
 		lli = ll_i2info(inode);
-		(*p)(env, cookie, "%lu/%u %o %u %d %p "DFID,
-		     inode->i_ino, inode->i_generation, inode->i_mode,
+		(*p)(env, cookie, "%llu/%u %o %u %d %p "DFID,
+		     (u64)inode->i_ino, inode->i_generation, inode->i_mode,
 		     inode->i_nlink, atomic_read(&inode->i_count),
 		     lli->lli_clob, PFID(&lli->lli_fid));
 	}
